fix(ic-icrc1-ledger): address PR #9799 review feedback - move types to icrc-ledger-types, use MTHD constants, check zero before conversion
